fix: tpm local repository change check added (#3)
* fix: check for repository change added before checkout

The Ansible git module will fail if there are any local changes in the
tpm repository, and there is no built in way to handle that gracefully.
This change adds a silent check for local diffs and if there are, skips
checking out the repository. The log output is hidden since in case there
is no repository the log is a wall of error text from git.

* chore: documentation